requirements.txt
# Geospatial & Data Pipeline
pystac-client>=0.7.0
stackstac>=0.4.3
geopandas>=0.13.0
rasterio>=1.3.0
shapely>=2.0.0
odc-stac>=0.3.6
planetary-computer>=1.0.0
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.5
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0

# API & Backend
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.22.0
streamlit-folium>=0.11.0
folium>=0.14.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.2.0
